+++ src/graph/core/DAGNode.java
@@ -25,7 +25,7 @@ import util.UtilityMethods;
  */
 public class DAGNode extends DAGObject implements Node {
 	/** The counter for assigning ids to edges. */
-	public static long idCounter_ = 1;
+	public static int idCounter_ = 1;
 
 	private static final long serialVersionUID = 2072866863770254720L;
 
@@ -52,7 +52,7 @@ public class DAGNode extends DAGObject implements Node {
 	}
 
 	@Override
-	protected long requestID() {
+	protected int requestID() {
 		return idCounter_++;
 	}
