+++ src/graph/core/cli/comparator/IDComparator.java
@@ -17,7 +17,7 @@ public class IDComparator extends DefaultComparator {
 	protected int compareInternal(Object o1, Object o2) {
 		// Need to be identifiable
 		if (o1 instanceof UniqueID && o2 instanceof UniqueID)
-			return Long.compare(((UniqueID) o1).getID(),
+			return Integer.compare(((UniqueID) o1).getID(),
 					((UniqueID) o2).getID());
 
 		return 0;
